Add files via upload
It is particularly important to describe the allocations that characterize or track the amounts of insurance claims. In doing so, it is important to distinguish between small claims and large claims. Minor indemnities are payments made by insurance companies which have the characteristic that each individual payment is negligible relative to the total indemnity, assuming that the number of payments is large. For such payments, distributions with a light (thin) tail are used. Large indemnities, on the other hand, imply payments whose value dominates the total amount of indemnity paid. For high-repayment modeling, appropriate distributions are with heavy (thick) tail.